Rename plugin to avoid conflict with elasticsearch plugin
When using the collectd elasticsearch plugin as a pip module, there
can be a import conflict with similarly named python modules on the
system. There is no ImportError as the module is found, instead,
collectd returns,

'Found a configuration for the "elasticsearch_collectd" plugin,
but the plugin isn't loaded or didn't register a configuration callback.'

This change renames the module to explicitly mention collectd and so
avoid conflicting names.